# Calculate psi values on a sparse grid, to accelerate multiple calculations.