--- a/src/sisl/CMakeLists.txt
+++ b/src/sisl/CMakeLists.txt
@@ -1,6 +1,17 @@
-foreach(source _indices _math_small _sparse_grid_ops)
+foreach(source _indices _math_small)
   add_cython_library(
     SOURCE ${source}.pyx
+    LIBRARY ${source}
+    OUTPUT ${source}_C
+    )
+  install(TARGETS ${source} LIBRARY
+    DESTINATION ${SKBUILD_PROJECT_NAME})
+endforeach()
+
+# Python files that can be compiled with cython (Pure Python syntax)
+foreach(source _sparse_grid_ops)
+  add_cython_library(
+    SOURCE ${source}.py
     LIBRARY ${source}
     OUTPUT ${source}_C
     )