--- a/src/sisl/physics/tests/test_density_matrix.py
+++ b/src/sisl/physics/tests/test_density_matrix.py
@@ -88,6 +88,11 @@ def setup():
     return t()
 
 
+@pytest.fixture(scope="module", params=["direct", "pre-compute"])
+def density_method(request):
+    return request.param
+
+
 @pytest.mark.physics
 @pytest.mark.densitymatrix
 class TestDensityMatrix:
@@ -183,14 +188,14 @@ class TestDensityMatrix:
             assert isinstance(BO, SparseOrbital)
             assert BO.shape[:2] == (D.geometry.no, D.geometry.no_s)
 
-    def test_rho1(self, setup):
+    def test_rho1(self, setup, density_method):
         D = setup.D.copy()
         D.construct(setup.func)
         grid = Grid(0.2, geometry=setup.D.geometry)
-        D.density(grid)
+        D.density(grid, method=density_method)
 
     @pytest.mark.filterwarnings("ignore", message="*is NOT Hermitian for on-site")
-    def test_rho2(self):
+    def test_rho2(self, density_method):
         bond = 1.42
         sq3h = 3.0**0.5 * 0.5
         lattice = Lattice(
@@ -214,23 +219,23 @@ class TestDensityMatrix:
         D = DensityMatrix(g)
         D.construct([[0.1, bond + 0.01], [1.0, 0.1]])
         grid = Grid(0.2, geometry=D.geometry)
-        D.density(grid)
+        D.density(grid, method=density_method)
 
         D = DensityMatrix(g, spin=Spin("P"))
         D.construct([[0.1, bond + 0.01], [(1.0, 0.5), (0.1, 0.1)]])
         grid = Grid(0.2, geometry=D.geometry)
-        D.density(grid)
-        D.density(grid, [1.0, -1])
-        D.density(grid, 0)
-        D.density(grid, 1)
+        D.density(grid, method=density_method)
+        D.density(grid, [1.0, -1], method=density_method)
+        D.density(grid, 0, method=density_method)
+        D.density(grid, 1, method=density_method)
 
         D = DensityMatrix(g, spin=Spin("NC"))
         D.construct(
             [[0.1, bond + 0.01], [(1.0, 0.5, 0.01, 0.01), (0.1, 0.1, 0.1, 0.1)]]
         )
         grid = Grid(0.2, geometry=D.geometry)
-        D.density(grid)
-        D.density(grid, [[1.0, 0.0], [0.0, -1]])
+        D.density(grid, method=density_method)
+        D.density(grid, [[1.0, 0.0], [0.0, -1]], method=density_method)
 
         D = DensityMatrix(g, spin=Spin("SO"))
         D.construct(
@@ -243,11 +248,11 @@ class TestDensityMatrix:
             ]
         )
         grid = Grid(0.2, geometry=D.geometry)
-        D.density(grid)
-        D.density(grid, [[1.0, 0.0], [0.0, -1]])
-        D.density(grid, Spin.X)
-        D.density(grid, Spin.Y)
-        D.density(grid, Spin.Z)
+        D.density(grid, method=density_method)
+        D.density(grid, [[1.0, 0.0], [0.0, -1]], method=density_method)
+        D.density(grid, Spin.X, method=density_method)
+        D.density(grid, Spin.Y, method=density_method)
+        D.density(grid, Spin.Z, method=density_method)
 
     @pytest.mark.filterwarnings("ignore", message="*is NOT Hermitian for on-site")
     def test_orbital_momentum(self):
@@ -474,20 +479,20 @@ class TestDensityMatrix:
         assert not np.allclose(D_mull, d_mull)
         assert np.allclose(D_mull[0], d_mull[0])
 
-    def test_rho_eta(self, setup):
+    def test_rho_eta(self, setup, density_method):
         D = setup.D.copy()
         D.construct(setup.func)
         grid = Grid(0.2, geometry=setup.D.geometry)
-        D.density(grid, eta=True)
+        D.density(grid, eta=True, method=density_method)
 
-    def test_rho_smaller_grid1(self, setup):
+    def test_rho_smaller_grid1(self, setup, density_method):
         D = setup.D.copy()
         D.construct(setup.func)
         lattice = setup.D.geometry.cell.copy() / 2
         grid = Grid(0.2, geometry=setup.D.geometry.copy(), lattice=lattice)
-        D.density(grid)
+        D.density(grid, method=density_method)
 
-    def test_rho_fail_p(self):
+    def test_rho_fail_p(self, density_method):
         bond = 1.42
         sq3h = 3.0**0.5 * 0.5
         lattice = Lattice(
@@ -513,9 +518,9 @@ class TestDensityMatrix:
         D.construct([[0.1, bond + 0.01], [(1.0, 0.5), (0.1, 0.1)]])
         grid = Grid(0.2, geometry=D.geometry)
         with pytest.raises(ValueError):
-            D.density(grid, [1.0, -1, 0.0])
+            D.density(grid, [1.0, -1, 0.0], method=density_method)
 
-    def test_rho_fail_nc(self):
+    def test_rho_fail_nc(self, density_method):
         bond = 1.42
         sq3h = 3.0**0.5 * 0.5
         lattice = Lattice(
@@ -543,7 +548,7 @@ class TestDensityMatrix:
         )
         grid = Grid(0.2, geometry=D.geometry)
         with pytest.raises(ValueError):
-            D.density(grid, [1.0, 0.0])
+            D.density(grid, [1.0, 0.0], method=density_method)
 
     def test_pickle(self, setup):
         import pickle as p