… src/sisl/_sparse_grid_ops.py
… src/sisl/_sparse_grid_ops.pyx