This PR addreses #493.

Concept and implementation details
---
The basic idea is to compute orbital values and store them in a sparse data structure. For now, the orbital values are computed in a new method -`Geometry.orbital_values`- and stored in a new class, `SparseGrid`. I didn't think too much on where to put the functionalities, so feel free to reorganize it however you think it fits best.

Once you have the `SparseGrid` data structure, you can call its `project_values` and `project_products` methods, which, through cython implementations, reduce the orbitals dimension and therefore project the values on a normal `sisl.Grid`. These routines accept coeficient arrays, which determine the weight of each value/product to be added according to the orbital indices from which it came. This generalization is then used to compute wavefunctions (using `project_values` with the coefficients being the eigenstate coefficients), and electronic density (using `project_products` with the coefficients being the density matrix). 

Some considerations about these two cython routines:
1. `project_products`: It loops over grid points to facilitate finding the overlaping orbitals. This makes it trivially easy to parallelize I would say, because there are no race conditions or any other problem. In principle, it should be as simple as changing the `range` call by cython's `prange` and compiling with openmp support, but I haven't tried.
2. `project_products` uses a dense coefficients matrix, because accessing a sparse one using indices is painfully slow if you put it inside a loop. Perhaps a smarter implementation could be added in the future to manage sparse matrices in case memory is a problem. But for now I don't see it as a problem, even for huge DFT systems.
3. `project_values`: In this case, there was no need to loop over some specific quantity, so it just loops over values and adds them as they come. This avoids the need for sorting the `SparseGrid`, but probably makes parallelization more difficult, since multiple iterations modify the same item of the grid. This is easy to solve by looping over grid points as in `project_products`, as long as the benefits are bigger than the cost of sorting.
4. Both: They use raveled grids and raveled indices, so they are not specific for 3D grids. This means that as long as you can generate a `SparseGrid` for something else (a path or a simple collection of points, then it would be called `SparseRealSpace` I guess), you can use these same routines out of the box. This opens the door for adding other `RealSpace` classes in `sisl`, instead of having to always go through grids, which sometimes are unnecessarily expensive for the goals of the user. 
5. In the same way, higher order products could be implemented. I don't know if they are useful for something.

Performance for computing the density
----

First, I needed to check that the implementation gives correct results. I took some SIESTA calculations that I had and compared the RHO file returned by SIESTA with the density calculated by this implementation using the DM file, in the exact same grid. Then took the absolute difference and computed histograms to check that the differences are close to 0. For both gamma point calculations and k-point calculations I get histograms like this:

![newplot - 2022-10-11T231952 096](https://user-images.githubusercontent.com/42074085/195200513-b72995d5-89a1-4f87-9054-05455850c8e1.png)

Which indicates that it is basically the same electronic density (differences < `7e-7`).

<details>
  <summary>Here's the code used to plot differences</summary>

  ```python
import sisl
import numpy as np
import plotly.express as px

def compare_siesta_sisl(fdf):
    
    DM = fdf.read_density_matrix()
    
    # RHO computed by SIESTA
    rho_SIESTA = fdf.read_grid("RHO")
    
    # Compute the density with sisl
    sisl_grid = DM.density(rho_SIESTA)
    
    # Compute a histogram of the difference
    counts, edges = np.histogram(abs((sisl_grid - rho_SIESTA)).grid, bins=100)

    # Plot it and show it
    px.bar(x=edges[:-1], y=counts / sisl_grid.size).update_layout(
        title="Histogram of difference between sisl and SIESTA RHO",
        xaxis_title="Difference", yaxis_title="Fraction of points"
    ).show()
    
fdf = sisl.get_sile("my.fdf")
compare_siesta_sisl(fdf) 
```
</details>

Now that I know that it works, I went on to check the performance. For this, I splitted the contributions of computing the orbital values and the density computation, given that one of the goals is to use the same orbital values to call `DM.density` multiple times. This is the scalability I get with grid points:

![newplot - 2022-10-11T234410 743](https://user-images.githubusercontent.com/42074085/195204097-256be3cb-4363-4cc4-adc3-0f390b5022f2.png)

So, both computations seem to scale close to linearly and are quite cheap, specially if we compare it with the current implementation, which I include here for reference using a logarithmic time scale:

![newplot - 2022-10-11T235418 188](https://user-images.githubusercontent.com/42074085/195205455-940b63b3-3a25-4362-ba21-e33ecf54379e.png)

For 1000 grid points in this system the current implementation takes 46s while the new one takes around 500ms, so an improvement of almost two orders of magnitude.

What is the source of this improvement? I found two main points:
- Of course, the implementation in cython of the loops that compute products makes it run much faster.
- But also, the way indices inside a sphere are retrieved in the current implementation (`grid.index(ia_atom.toSphere(ia_xyz))`) seems to be very inneficient. By changing it to the approach in the new implementation (getting a bounding box and then filtering by radius) I was able to improve the performance of the current implementation from 46s to 12s. The new method is probably a bit worse for skewed grids, where the bounding box is overestimated, but still I expect that it should be better.
- Finally, I think getting values of the sparse DM by querying with indices is probably also slowing down the current implementation.

<details>
  <summary>Here's the code used to plot the scalability</summary>

```python
import sisl
import numpy as np
import time
import plotly.graph_objs as go

def time_density(DM, grid_shape):
    geometry = DM.geometry
    
    # Initialize grid
    grid = sisl.Grid(grid_shape, geometry=geometry)
    
    # Compute orbital values and time it
    start_psi = time.time()
    psi_values = geometry.orbital_values(grid.shape)
    end_psi = time.time()
    
    # Use orbital values to compute density and time it
    start_dens = time.time()
    DM.density(grid, psi_values=psi_values)
    end_dens = time.time()
    
    # Return both times.
    return (end_psi - start_psi, end_dens - start_dens)

def scalability_density(fdf, ref_old=False):
    
    DM = fdf.read_density_matrix()
    
    sizes = []
    psi_times = []
    dens_times = []

    for shape in np.arange(5, 50, 5):
        psi_time, dens_time = time_density(DM, grid_shape=(shape, shape, shape))

        sizes.append(shape**3)
        psi_times.append(psi_time)
        dens_times.append(dens_time)
    
    if ref_old:
        # Get the old performance for a coarse grid, for reference
        old_grid_shape = (10,10,10)
        old_grid = sisl.Grid(old_grid_shape, geometry=DM.geometry)

        start = time.time()
        DM.old_density(old_grid)
        end = time.time()

        old_time = end - start
        old_size = np.prod(old_grid_shape)
    
    fig = go.Figure()
    
    fig.add_scatter(
        x=sizes, y=np.array(psi_times) + dens_times, name="Total"
    ).add_scatter(
        x=sizes, y=dens_times, name="Density"
    ).add_scatter(
        x=sizes, y=psi_times, name="Orbitals on grid"
    ).update_layout(
        xaxis_title="Number of grid points",
        yaxis_title="Elapsed time (s)",
        title="Scaling with grid size",
        yaxis_showgrid=True, xaxis_showgrid=True
    )
    
    if ref_old:
        fig.add_scatter(
            x=[old_size], y=[old_time], name="Previous implementation"
        )
        fig.update_layout(yaxis_type="log", yaxis_title="log(Elapsed time) (s)")
    
    fig.show()
    
    return fig

fdf = sisl.get_sile("./Final_coor/mdsamp-2.0V/1000/Au_Water_MD.fdf")
fig = scalability_density(fdf)
```
</details>

Performance for computing wavefunctions
----

For wavefunctions, I also checked with the differences histogram that I get the same results as with the current implementation for gamma point and k-point calculations.

The performance of the current implementation and the implementation through stored psi values is very similar (the new implementation is slightly more expensive, about 10%). This is to be expected, since the expensive part is computing the orbital values, which is done in a similar way in both cases. However the new implementation uses much more memory, since it stores the values instead of adding them on the fly. This is why I kept the current implementation as the default. Computing the wavefunction through stored psi values should only be done when you are computing multiple wavefunctions, or you have already computed the psi values for something else. If that is the case, calculating wavefunctions is of course very cheap in terms of time.

<details>
  <summary>Here's the code used to compare wavefunctions with both methods</summary>

```python
import sisl
import numpy as np
import plotly.express as px

def compare_wf_strategies(H, k, grid_shape):
    
    geometry = H.geometry
    geometry.xyz = geometry.translate2uc().xyz
    
    eig = H.eigenstate(k=k)[1]

    wf_grid = sisl.Grid(grid_shape, geometry=geometry, dtype=eig.dtype)
    eig.wavefunction(wf_grid)
    
    psi_values = geometry.orbital_values(wf_grid.shape, pbc=True)
    new_wf_grid = eig.wavefunction(wf_grid, psi_values=psi_values)
    
    # Compute a histogram of the difference
    counts, edges = np.histogram(abs(new_wf_grid - wf_grid).grid, bins=100)

    # Plot it and show it
    px.bar(x=edges[:-1], y=counts / wf_grid.size).update_layout(
        title="Histogram of difference between different wavefunction strategies",
        xaxis_title="Difference", yaxis_title="Fraction of points"
    ).show()

fdf = sisl.get_sile("my.fdf")
compare_wf_strategies(fdf.read_hamiltonian(), k=(0.5,0,0), grid_shape=(10, 10, 10))  
```
</details>

Looking forward to your thoughts on this!