@pfebrer 2022-10-12
By the way you need cython 3.0 to compile it, which is still an alpha version, so you can install it with `pip install --pre Cython`