@zerothi 2022-10-12
> By the way you need cython 3.0 to compile it, which is still an alpha version, so you can install it with `pip install --pre Cython`

I'll have a look! Thanks for this.

I am not too happy about the alpha dependency... What feature is it that you need? Can we use older cython and update when 3 comes out?