@pfebrer 2022-10-12
Cython 3.0 has been in alpha for like 2 years I think and it is now in its 11th iteration. I think it is pretty safe to use. Not using it is like using siesta 4.0 instead of 4.1b4 Anyway this is not a dependency that you will impose on users, since you will provide the sources already compiled. So I don't think it's that dramatic.

The problem is that fused types in the pure python syntax don't work in 0.29, so I would have to rewrite everything with cython syntax.