@pfebrer 2022-10-13
Ok, it now works with 0.29.32. I didn't know you could mix both sintaxes. Since you can, this is an easy fix that doesn't annoy me too much and is easy to revert.

I have also created an issue in cython to see if they can solve it in 0.29: https://github.com/cython/cython/issues/5085