@pfebrer 2022-11-13
I've done some changes. I realized that if I stored the sparse grid as a CSR matrix where rows are the elements of the grid (if it is a 3D grid, it has been raveled to 1D) and the columns are the extra dimension it would be benefitial because:

- I don't have to create and maintain a new sparse structure.
- I can store orbitals in columns by their sc index, following the conventions in the rest of the code. 
- It requires less memory than my previous implementation. Because of this it also got a little faster.
- **Projecting the wavefunctions now becomes a simple matrix multiplication** (psi X state), where state can be 2D resulting in 1D grid elements (as opposed to 0D or scalar).

The structure is cleaner now, so generic routines can be implemented for useful things and they don't look ultra-complex. For example I implemented:
- Orbital products can take a sisl sparse matrix with `dim > 1` and perform all orbital products in the same go, resulting in 1D grid elements. This is currently implemented by creating a dense matrix containing the pointers to nonzero data, so that the access can be as fast as possible. I also have a not so slow algorithm for avoiding creating a dense array (which is not in the PR, yet), but is still much slower.
- **Reduction of the grid dimensionality while projecting the values** to save memory. This plays nicely with the 1D grid elements. For example, I wanted to reduce the X and Y axes to keep only Z. If I can't reduce them on the fly, the memory requirements blow up if I have 1D grid elements, but they are below those of a regular 3D grid if I use on the fly reduction.

I also added lots of tests, so I now feel much more confident about its correctness and it will be safer to do modifications on it.

There is one problem though: I don't know how to make the functionality here play nicely with the current API for `density` and `wavefunction`. The current API accepts a grid that modifies in place. However, if you have stored the psi values already in a `SparseGrid4D`, to me it makes more sense that this grid has a method, e.g. `reduce_orbitals`, that returns a new grid. I would like to hear your thoughts on how would you manage this. Also, because ND grids still do not exist in `sisl`, when multiple coefficients are computed I need to return a `numpy` array.

Finally here are some plots of using multiple density matrices in the same product computation vs using them one by one. The plots are for different grid shapes, for a system with 3000 orbitals:

![grid_10_10_10](https://user-images.githubusercontent.com/42074085/201540089-d92e47d6-39de-43c0-aaae-921fade928e6.png)

![grid_30_30_600_larger](https://user-images.githubusercontent.com/42074085/201540126-958cf865-415d-44c5-9def-8276a5b79924.png)

Overall I think it's very nice because it has allowed me to calculate `LDOS(E)` maps for huge systems and 240 energy points in less than 3 minutes, and before this it took around 4-5h :) This could enable an LDOS method in `tbtncSile`.