@pfebrer 2023-08-13
If this went in, I'm not sure it is more expensive to compute the LDOS directly by squaring wavefunctions. It would be something like this:

```math
 \textbf{LDOS}(x, E) = ( \textbf{V}  \textbf{C})^{2_{el-wise}}\textbf{N}
```

Where:
 - `C` is the eigenstate coefficients matrix (orbitals X wavefunctions).
 - `V` is the sparse matrix implemented in this PR storing the orbital values (point grids X orbitals).
 - `2_el-wise` represents the element-wise squaring of the `VC` matrix (to get the squared wavefunctions).
 - `N` is the occupations matrix (wavefunctions X energies).

`V` and `N` are sparse, but `C` and `VC` are not.
 
You end up with a matrix LDOS (point grids X energies). I'm not sure this is more expensive than the DM way, specially in python since these are simple matrix multiplications that are already implemented in `scipy`/`numpy` and which are probably already implemented in GPUs, versus the custom code for finding products of orbitals that you need to compute the LDOS from the DM. In terms of memory, I don't know what's more expensive.

Also, C can contain a third dimension for the k points, and doing the same operation you end up with `LDOS(x, E, k)`. I'm sure this is an operation GPUs would love doing :)