@pfebrer 2023-11-08
I noticed that $V^TV$ ($V$ being the matrix of psi values) was the overlap matrix.

I know this is not the most efficient way, it is faster use fourier transforms. But it was easy to add and since there is no way currently of computing the overlap in sisl I just added it here :sweat_smile: